+++ src/services/api/libs/auth/user.ts
@@ -1,15 +1,10 @@
-import { response } from "../../libs/handler";
-import { DynamoDBClient } from "@aws-sdk/client-dynamodb";
-import { SeatoolService } from "../../services/seatoolService";
-import { APIGatewayEvent } from "aws-lambda";
 import {
   CognitoIdentityProviderClient,
   ListUsersCommand,
   UserType as CognitoUserType,
 } from "@aws-sdk/client-cognito-identity-provider";
 import { CognitoUserAttributes } from "shared-types";
-
-const dynamoInstance = new DynamoDBClient({ region: process.env.region });
+import { APIGatewayEvent } from "aws-lambda";
 
 export function getAuthDetails(event: APIGatewayEvent) {
   try {
@@ -53,7 +48,7 @@ export const getParsedObject = (obj: CognitoUserAttributes) =>
     })
   );
 
-async function lookupUserAttributes(
+export async function lookupUserAttributes(
   userId: string,
   poolId: string
 ): Promise<CognitoUserAttributes> {
@@ -89,43 +84,3 @@ async function fetchUserFromCognito(userID: string, poolID: string) {
   const currentUser = listUsersResponse.Users[0];
   return currentUser;
 }
-
-export const getSeatoolData = async (event: APIGatewayEvent) => {
-  try {
-    const authDetails = getAuthDetails(event);
-    const userAttributes = await lookupUserAttributes(
-      authDetails.userId,
-      authDetails.poolId
-    );
-
-    const stateCode = event.pathParameters.stateCode;
-
-    if (
-      !userAttributes ||
-      !userAttributes["custom:state_codes"].includes(stateCode)
-    ) {
-      return response({
-        statusCode: 401,
-        body: { message: "User is not authorized to access this resource" },
-      });
-    }
-
-    const seaData = await new SeatoolService(dynamoInstance).getIssues({
-      tableName: process.env.tableName,
-      stateCode: stateCode,
-    });
-
-    return response<unknown>({
-      statusCode: 200,
-      body: seaData,
-    });
-  } catch (error) {
-    console.error({ error });
-    return response({
-      statusCode: 404,
-      body: { message: error },
-    });
-  }
-};
-
-export const handler = getSeatoolData;
